+++ doc/releases/changelog-dev.md
@@ -797,7 +797,6 @@
   [(#9702)](https://github.com/PennyLaneAI/pennylane/pull/9702)
   [(#9723)](https://github.com/PennyLaneAI/pennylane/pull/9723)
   [(#9729)](https://github.com/PennyLaneAI/pennylane/pull/9729)
-  [(#9744)](https://github.com/PennyLaneAI/pennylane/pull/9744)
   [(#9737)](https://github.com/PennyLaneAI/pennylane/pull/9737)
   [(#9730)](https://github.com/PennyLaneAI/pennylane/pull/9730)
