+++ tests/ops/op_math/test_controlled2.py
@@ -19,13 +19,12 @@ import numpy as np
 import pytest
 
 import pennylane as qp
-from pennylane.core import Operator2
 from pennylane.ops.op_math.controlled import Controlled
 from pennylane.ops.op_math.controlled2 import Controlled2, ControlledOp2
-from pennylane.typing import Bool, Float, Wire
 from pennylane.wires import Wires
+from tests.core.operator.operator2_utils import OneWireDynOp
 
-# pylint: disable=unused-argument,too-few-public-methods,useless-parent-delegation
+# pylint: disable=unused-argument,too-few-public-methods
 
 
 class TestControlled2:
@@ -249,18 +248,6 @@ class TestControlled2:
         qp.assert_equal(simplified_op, qp.ctrl(qp.MultiRZ(0.5, [0, 1, 2]), control=[3, 4, 5]))
 
 
-class CustomOp(Operator2):
-
-    dynamic_argnames = ("theta",)
-
-    wire_argnames = ("wires",)
-
-    arg_specs = {"theta": Float, "wires": Wire[1]}
-
-    def __init__(self, theta, wires):
-        super().__init__(theta, wires)
-
-
 class TestControlledOp2:
     """Tests the ControlledOp2 class."""
 
@@ -414,19 +401,9 @@ class TestControlledOp2:
         generator = qp.Projector([1], wires=1) @ qp.Hamiltonian([-0.5], [qp.PauliX(0)])
         qp.assert_equal(op.generator(), generator)
 
-    def test_create_abstract_op(self):
-        """Tests creating an abstract operator."""
-
-        op = ControlledOp2(CustomOp, Wire[2])
-        assert op.control_wires == Wire[2]
-        assert op.target_wires == Wire[1]
-        assert op.control_values == Bool[2]
-        assert op.work_wires == Wire[0]
-        assert op.wires == Wire[3]
-
     def test_create_controlled_op2(self):
         """Tests qp.ctrl on Operator2 creates a ControlledOp2."""
 
-        op = CustomOp(0.5, wires=[0])
-        op = qp.ctrl(CustomOp(0.5, wires=[0]), control=[1], control_values=0)
+        op = OneWireDynOp(0.5, wires=[0])
+        op = qp.ctrl(OneWireDynOp(0.5, wires=[0]), control=[1], control_values=0)
         assert isinstance(op, ControlledOp2)
